refactors taggable behaviour
